리뷰 글삭제 servlet 수정
- 본인 글만 수정할 수 있는 변별 코드 삽입

+++ semi/src/main/java/semi/servlet/review/ReviewDeleteServlet.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import semi.beans.ReviewDao;
+import semi.beans.ReviewDto;
 
 @WebServlet(urlPatterns="/review/delete.kh")
 public class ReviewDeleteServlet extends HttpServlet {
@@ -16,9 +17,21 @@ public class ReviewDeleteServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		try {
+			req.setCharacterEncoding("UTF-8");
+			String memberId = (String)req.getSession().getAttribute("login");
+			
 			int reviewNo = Integer.parseInt(req.getParameter("reviewNo"));
 			
 			ReviewDao reviewDao = new ReviewDao();
+			ReviewDto reviewDto = new ReviewDto();
+			reviewDto = reviewDao.showDetail(reviewNo);
+			boolean isOwner = memberId.equals(reviewDto.getReviewMemberId());
+			
+			if(!isOwner) {
+				resp.sendRedirect("detail.jsp?reviewNo="+reviewNo);
+				return;
+			}
+			
 			boolean success = reviewDao.delete(reviewNo);
 			
 			if(success) {
